requirements.txt
python-telegram-bot==21.0.1
yt-dlp
psycopg2-binary
python-dotenv
matplotlib
reportlab
aiohttp
aiohttp-retry
aiofiles
fastapi
uvicorn

# shazamio — Python 3.12 va pastroqda ishlaydi (3.13+ da audioop olib tashlangan)
# shazamio
